feature(ConvertFlow): Refactoring code
