+++ integrations/ConvertFlow/browser.js
@@ -3,8 +3,6 @@
 import { NAME } from "./constants";
 import logger from "../../utils/logUtil";
 import ScriptLoader from "../ScriptLoader";
-import { getHashFromArray } from "../utils/commonUtils";
-import makeACall from "./utils";
 
 class ConvertFlow {
   constructor(config) {
@@ -12,10 +10,6 @@ class ConvertFlow {
     this.toggleToSendData = config.toggleToSendData;
     this.eventsList = config.eventsList;
     this.eventsMappping = config.eventsMappping;
-    this.blacklistedEvents = config.blacklistedEvents;
-    this.whitelistedEvents = config.whitelistedEvents;
-    this.oneTrustCookieCategories = config.oneTrustCookieCategories;
-    this.eventFilteringOption = config.eventFilteringOption;
     this.name = NAME;
   }
 
@@ -45,30 +39,11 @@ class ConvertFlow {
     logger.debug("===In convertflow Identify===");
     const { message } = rudderElement;
     const email = message.context.traits?.email || message.traits?.email;
-    if (email) {
-      const payload = { email, override: true };
-      window.convertflow.identify(payload);
+    if (!email) {
+      logger.error("email is rquired for identify call");
     }
-  }
-
-  trigger() {
-    const standardEventsMap = getHashFromArray(this.eventsMappping);
-    const standardEventsList = [
-      "cfReady",
-      "cfView",
-      "cfConversion",
-      "cfCompletion",
-      "cfSubmit",
-      "cfAddToCart",
-      "cfClosed",
-    ];
-    standardEventsList.forEach((events) => {
-      window.addEventListener(events, function (event, data) {
-        if (this.eventsList.includes(events)) {
-          makeACall(standardEventsMap, events, data);
-        }
-      });
-    });
+    const payload = { email, override: true };
+    window.convertflow.identify(payload);
   }
 }
